requirements.txt
PyMuPDf
Pillow
google.generativeai
dotenv